refactor: align lobby package to api application domain structure
