+++ src/main/java/com/codenames/codenames/backend/lobby/api/dto/LobbyResponse.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames.backend.lobby.dto;
+package com.codenames.codenames.backend.lobby.api.dto;
 
 import java.util.List;
