+++ src/main/java/com/codenames/codenames/backend/recovery/snapshot/SystemSnapshot.java
@@ -1,6 +1,6 @@
 package com.codenames.codenames.backend.recovery.snapshot;
 
-import com.codenames.codenames.backend.lobby.dto.PlayerDto;
+import com.codenames.codenames.backend.lobby.api.dto.PlayerDto;
 import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import java.util.List;
 import java.util.Map;
