+++ src/test/java/com/codenames/codenames/backend/lobby/api/LobbySocketControllerTest.java
@@ -12,7 +12,7 @@ import static org.mockito.Mockito.when;
 import com.codenames.codenames.backend.game.application.GameService;
 import com.codenames.codenames.backend.lobby.api.dto.JoinMessage;
 import com.codenames.codenames.backend.lobby.domain.Player;
-import com.codenames.codenames.backend.lobby.services.LobbyService;
+import com.codenames.codenames.backend.lobby.application.LobbyService;
 import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import com.codenames.codenames.backend.shared.websocket.SessionRegistry;
 import java.util.List;
