+++ src/test/java/com/codenames/codenames/backend/recovery/SystemStatePersistenceServiceTest.java
@@ -7,8 +7,8 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.codenames.codenames.backend.game.application.GameService;
-import com.codenames.codenames.backend.lobby.dto.PlayerDto;
-import com.codenames.codenames.backend.lobby.services.LobbyService;
+import com.codenames.codenames.backend.lobby.api.dto.PlayerDto;
+import com.codenames.codenames.backend.lobby.application.LobbyService;
 import com.codenames.codenames.backend.recovery.snapshot.SystemSnapshot;
 import com.codenames.codenames.backend.game.dto.GameStateDataTransferObject;
 import java.util.List;
