+++ src/test/java/com/codenames/codenames/backend/recovery/SystemStateRecoveryServiceTest.java
@@ -12,10 +12,10 @@ import com.codenames.codenames.backend.game.application.GameService;
 import com.codenames.codenames.backend.game.domain.GameManager;
 import com.codenames.codenames.backend.game.domain.GameManagerFactory;
 import com.codenames.codenames.backend.game.dto.ClueDto;
-import com.codenames.codenames.backend.lobby.Lobby;
-import com.codenames.codenames.backend.lobby.dto.PlayerDto;
-import com.codenames.codenames.backend.lobby.services.LobbyCodeGenerator;
-import com.codenames.codenames.backend.lobby.services.LobbyService;
+import com.codenames.codenames.backend.lobby.domain.Lobby;
+import com.codenames.codenames.backend.lobby.api.dto.PlayerDto;
+import com.codenames.codenames.backend.lobby.application.LobbyCodeGenerator;
+import com.codenames.codenames.backend.lobby.application.LobbyService;
 import com.codenames.codenames.backend.recovery.snapshot.SystemSnapshot;
 import com.codenames.codenames.backend.game.dto.CardDataTransferObject;
 import com.codenames.codenames.backend.game.mapping.DataTransferObjectService;
